requirements.txt
python-telegram-bot==26.0
requests>=2.28.0
aiosqlite>=0.19.0
